Refactor flow canvas into modular hooks and components
