+++ src/hooks/useLatest.ts
@@ -1,0 +1,11 @@
+import { useEffect, useRef } from 'react';
+
+const useLatest = <T,>(value: T) => {
+  const ref = useRef(value);
+  useEffect(() => {
+    ref.current = value;
+  }, [value]);
+  return ref;
+};
+
+export default useLatest;
