+++ src/utils/array.ts
@@ -1,0 +1,1 @@
+export const uniqueList = (values: string[]) => Array.from(new Set(values));
